Move create window code to libopenrct2ui
